requirements.txt
google-api-python-client==2.149.0
python-dotenv==1.0.1
streamlit==1.40.2
youtube-transcript-api==0.6.2
pandas==2.2.3
yt-dlp==2025.11.12
